Use separate model for new STV form
