+++ src/components/AdminAdvice/AdminAdvice.less
@@ -1,0 +1,30 @@
+/*
+ * Chron.
+ * Copyright (c) 2020 Alisa Belyaeva, Ata Ali Kilicli, Amaury Martiny,
+ * Daniil Mordasov, Liam O’Flynn, Mikhail Orlov.
+ * -----
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * -----
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * -----
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+ @import '../../styles/colors.less';
+
+ .admin-advice {
+  display: flex;
+  background-color: rgba(@panel, 0.5);
+  border-radius: 4px;
+  padding: .5rem;
+  justify-content: space-between;
+  align-items: center;
+  margin: 0.25rem;
+  min-height: 3rem;
+}
