+++ src/containers/AdminSTV/AdminSTV.less
@@ -46,14 +46,3 @@
   align-items: center;
   justify-content: space-around;
 }
-
-.admin-advice {
-  display: flex;
-  background-color: rgba(@panel, 0.5);
-  border-radius: 4px;
-  padding: .5rem;
-  justify-content: space-between;
-  align-items: center;
-  margin: 0.25rem;
-  min-height: 3rem;
-}
